+++ src/components/Link/tests/Link.test.stories.tsx
@@ -1,0 +1,111 @@
+import { StoryObj } from "@storybook/react";
+import { userEvent, expect, fn } from "@storybook/test";
+import Link from "components/Link";
+
+export default {
+	title: "Components/Link/tests",
+	component: Link,
+	parameters: {
+		iframe: {
+			url: "https://reshaped.so/docs/components/breadcrumbs",
+		},
+	},
+};
+
+export const children: StoryObj = {
+	name: "children",
+	render: () => <Link>Trigger</Link>,
+	play: async ({ canvas }) => {
+		const el = canvas.getByText("Trigger");
+
+		expect(el).toBeInTheDocument();
+		expect(el.tagName).toBe("SPAN");
+	},
+};
+
+export const href: StoryObj = {
+	name: "href",
+	render: () => <Link href="https://reshaped.so">Trigger</Link>,
+	play: async ({ canvas }) => {
+		const el = canvas.getByRole("link");
+
+		expect(el).toHaveAttribute("href", "https://reshaped.so");
+	},
+};
+
+export const onClick: StoryObj<{ handleClick: ReturnType<typeof fn> }> = {
+	name: "onClick",
+	args: {
+		handleClick: fn(),
+	},
+	render: (args) => <Link onClick={args.handleClick}>Trigger</Link>,
+	play: async ({ canvas, args }) => {
+		const { handleClick } = args;
+		const el = canvas.getAllByRole("button")[0];
+
+		await userEvent.click(el);
+
+		expect(el).toHaveAttribute("type", "button");
+		expect(handleClick).toHaveBeenCalledTimes(1);
+		expect(handleClick).toHaveBeenCalledWith(expect.objectContaining({ target: el }));
+	},
+};
+
+export const hrefOnClick: StoryObj<{ handleClick: ReturnType<typeof fn> }> = {
+	name: "href + onClick",
+	args: {
+		handleClick: fn(),
+	},
+	render: (args) => (
+		<Link
+			onClick={(e) => {
+				e.preventDefault();
+				args.handleClick(e);
+			}}
+			href="https://reshaped.so"
+		>
+			Trigger
+		</Link>
+	),
+	play: async ({ canvas, args }) => {
+		const { handleClick } = args;
+		const el = canvas.getByRole("link");
+
+		await userEvent.click(el);
+
+		expect(el).toHaveAttribute("href", "https://reshaped.so");
+		expect(handleClick).toHaveBeenCalledTimes(1);
+		expect(handleClick).toHaveBeenCalledWith(expect.objectContaining({ target: el }));
+	},
+};
+
+export const disabled: StoryObj = {
+	name: "disabled",
+	render: () => (
+		<Link disabled onClick={() => {}}>
+			Trigger
+		</Link>
+	),
+	play: async ({ canvas }) => {
+		const el = canvas.getAllByRole("button")[0];
+
+		expect(el).toBeDisabled();
+	},
+};
+
+export const className: StoryObj = {
+	name: "className, attributes",
+	render: () => (
+		<div data-testid="root">
+			<Link className="test-classname" attributes={{ id: "test-id" }}>
+				Trigger
+			</Link>
+		</div>
+	),
+	play: async ({ canvas }) => {
+		const root = canvas.getByTestId("root").firstChild;
+
+		expect(root).toHaveClass("test-classname");
+		expect(root).toHaveAttribute("id", "test-id");
+	},
+};