+++ src/components/Loader/Loader.tsx
@@ -1,11 +1,10 @@
-import React from "react";
 import { classNames, responsiveClassNames } from "utilities/helpers";
 import type * as T from "./Loader.types";
 import s from "./Loader.module.css";
 
 const Loader = (props: T.Props) => {
 	const { size = "small", color = "primary", className, attributes } = props;
-	const ariaLabel = attributes?.["aria-label"] as React.AriaAttributes["aria-label"];
+	const ariaLabel = props.ariaLabel || attributes?.["aria-label"];
 	const rootClassNames = classNames(
 		s.root,
 		className,